+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/setup/state/ClientMutableState.kt
@@ -42,5 +42,5 @@ internal interface ClientMutableState : ClientState {
     /**
      * Clears the state of [ClientMutableState].
      */
-    override fun clearState()
+    public fun clearState()
 }