+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/Mother.kt
@@ -127,10 +127,6 @@ internal object Mother {
             override val isInitialized: Boolean = true
 
             override val isNetworkAvailable: Boolean = true
-
-            override fun clearState() {
-                // Nothing to do
-            }
         }
     }